Implement Hello World compilation support
Example programs which can be compiled can be found in the doc/example
folder.

A contemporary version of libbsa is required for this program (bscc)
to succeed at compilation, as well as to run the compiled Hello World
application.

Important short term TODOs are improved validation, and a test suite.