Upgrade to Cabal v1.18, fixing ``cabal sdist''
Previously we used Cabal v1.16.  Because we have a postSDist hook, the
standard ``cabal sdist'' command didn't work.  There was however a
workaround; using ``./Setup.hs sdist'' performed the sdist and invoked
our hook.

The upgrade to Cabal v1.18 fixes
<https://github.com/haskell/cabal/issues/403>, making ``cabal sdist''
invoke our hook.

There is a change in behaviour, so we have to update our hook for it
to continue to work.  In particular, previously the hook was invoked
after the tarball had been created.  With ``cabal sdist'' in Cabal
v1.18, our postSDist hook is invoked before the files are packed up
into a tarball.