Replace unnecessary tests of IO actions
We no longer need to test the IO actions because the IO actions are
trivially implemented in terms of pure functions.  That's good because
our tests of the IO actions were ugly: they had to redirect stdout.

Said pure functions themselves are pretty trivial, so there's not much
point in testing them either.  Nonetheless, we those functions just to
maintain a working test infrastructure.