Remove all external dependencies of HelpAndVersion
Suppose you build bscc in a Cabal sandbox.  bscc-help2man-shim, which
uses this module, will be invoked.  Due to the way it's invoked, it
will not have access to the modules in the sandbox.  Hence, in order
to not complicate the installation instructions needlessly, the shim
should only depend upon base.