Relax upper bound on base dependency
Doing so goes against the recommendation of the Haskell Package
Versioning Policy.  Nonetheless, I think it's a good idea.  I sampled
some important packages and they all do the same.  And there's a good
reason.  We want to be able to compile on future versions of GHC, such
as v7.8.