Add a dependency on lens
Lens is a state of the art lenses package that I would like to better
learn.  Already I can tell that its uniplate support is awesome.

What is also nice about the library is that a lot of it is
mathematically principled.  Good laws for types make it hard to
accidentally write incorrect code.