Use strongly typed filepaths
These come form the pathtype package.  It is nice in particular to be
able to distinguish between relative and absolute paths.

However, we still use Prelude's FilePath in Main to represent a
FilePaths from the command line which could be absolute or relative.
This is easier than using an existential mostly-private class from
pathtype to achieve the same thing.