Remove comment saying Proc shouldn't have Eq class
The comment had claimed Proc shouldn't be an instance of Eq because
two (physically separate) procedures can have the same name in
different modules.  However, that was a misunderstanding of Eq.  For
Eq for Proc we should consider its tree data structure.  It doesn't
mind that we might interpret equal Procs differently in different
contexts.