Specify a precise version of the Cabal dependency
Previously we used a range.  However, that causes the build to fail if
uuagc-cabal and our Setup.hs try to use different versions.

In fact, unfortunately it seems I have to specify a development
release of Cabal; only by using a development version can I seem to
stop the build errors for now.  This is not ideal, nor is it a long
term solution.

This is presumably related to
<https://github.com/haskell/cabal/issues/948>.