Generate an LLVM AST for the IR
Previously we directly generated a String of LLVM Assembly for the
Intermediate Representation.  This was hugely suboptimal, at least
theoretically.  For a start, String is not packed like Text is.
However it was worse than that.  We were appending Strings without
considering associativity, because who wants to have to consider the
associativity?  Unfortunately, left associative (++) is quadratic.

That rationale justifies changing the implementation.  However, it
probably suggest generating a ``Seq Text'' instead of a ``String''. It
doesn't obviously suggest using llvm-general{-pure,} to generate an
AST.  I could say something about how generating the AST is slightly
more typesafe; that is definitely true, given the number of type
errors I had to fix whilst developing this change.  But to be honest,
my main motivation was simply that I wanted to give llvm-general a
shot.