Use monoids for command line options
This is based upon a observation that Duncan Coutts made for Cabal, at
<http://www.haskell.org/pipermail/cabal-devel/2007-December/001509.html>.
Namely, any command line option can be modelled as a list or a Last
monoid.

This change uses a more-exotic extension than is necessary.  Use GADTs
to implement an existential type, not the ExistentialTypes extension.
This is because I find the GADT syntax for existentials more
comprehensible.