Release new version, v0.1.2
This seems like a sensible time to increase the version number.  This
is because for the first time the existing functionality, whilst
limited, has no known major issues.  Famous last words, I know, but
recently we added some good error handling to the semantic analysis.

Of note I'm not bumping the version number of bscc-setuphs.  There's
no point because it hasn't changed.